requirements.txt
numpy
pandas
joblib
requests
streamlit
tensorflow
scikit-learn
python-dotenv
streamlit-lottie
